requirements.txt
poetry
nougat-ocr
transformers==4.38.2